requirements.txt
streamlit==1.18.1
os==1.0.0
python-dotenv==0.21.0
groq==1.0.0
base64==1.3.0
re==2.3.1
streamlit-chat==0.1.0
pygments==2.13.1